slight modification

+++ Project 3/heap.h
@@ -2,7 +2,7 @@
 
 // struct elements for the head
 struct Element {
-    int u;
+    int u; // points to vertex V[vertex]
     float key;
 };
 
@@ -23,7 +23,7 @@ public:
     void buildHeap(Element *arr, int size);
     void insert(int u, float key, int &index, int flagFind);
     int extractMin(int flagFind);
-    void decreaseKey(int index, int newKey);
+    void decreaseKey(int index, float newKey);
     void minHeapify(int index);
     void printHeap();
     void swap(int x, int y);
